+++ src/main/java/database/OrderDAO.java
@@ -1,10 +1,15 @@
 package database;
 
 import model.*;
+import util.Hash;
+import util.RSA;
 
 import javax.servlet.http.HttpServletRequest;
+import java.nio.charset.StandardCharsets;
+import java.security.NoSuchAlgorithmException;
 import java.sql.*;
 import java.sql.Date;
+import java.text.SimpleDateFormat;
 import java.time.LocalDateTime;
 import java.util.*;
 
@@ -56,6 +61,8 @@ public class OrderDAO extends AbsDAO<Order>{
 
                 // Áp dụng phương thức kiểm tra và cập nhật trạng thái chữ ký
                 statusSignature = checkAndUpdateSignatureStatus(con, idImport, statusSignature);
+                statusSignature = checkChangeDatabase(con, idUser, idImport, statusSignature);
+
 
                 User u = new UserDAO().selectById(idUser);
                 Payment pay = new PaymentDAO().selectById(idPayment);
@@ -100,6 +107,8 @@ public class OrderDAO extends AbsDAO<Order>{
 
                 // Áp dụng phương thức kiểm tra và cập nhật trạng thái chữ ký
                 statusSignature = checkAndUpdateSignatureStatus(con, idImport, statusSignature);
+                statusSignature = checkChangeDatabase(con, idUser, idImport, statusSignature);
+
 
 
                 User u = new UserDAO().selectById(idUser);
@@ -249,6 +258,8 @@ public class OrderDAO extends AbsDAO<Order>{
 
                 // Áp dụng phương thức kiểm tra và cập nhật trạng thái chữ ký
                 statusSignature = checkAndUpdateSignatureStatus(con, idImport, statusSignature);
+                statusSignature = checkChangeDatabase(con, idUser, idImport, statusSignature);
+
 
                 User u = new UserDAO().selectById(idUser);
                 Payment pay = new PaymentDAO().selectById(idPayment);
@@ -309,6 +320,7 @@ public class OrderDAO extends AbsDAO<Order>{
 
                 // Áp dụng phương thức kiểm tra và cập nhật trạng thái chữ ký
                 statusSignature = checkAndUpdateSignatureStatus(con, idImport, statusSignature);
+                statusSignature = checkChangeDatabase(con, idUser, idImport, statusSignature);
 
 
                 User u = new UserDAO().selectById(idUser);
@@ -356,6 +368,7 @@ public class OrderDAO extends AbsDAO<Order>{
 
                 // Áp dụng phương thức kiểm tra và cập nhật trạng thái chữ ký
                 statusSignature = checkAndUpdateSignatureStatus(con, idImport, statusSignature);
+                statusSignature = checkChangeDatabase(con, userId, idImport, statusSignature);
 
 
                 User u = new UserDAO().selectById(idUser);
@@ -405,6 +418,7 @@ public class OrderDAO extends AbsDAO<Order>{
 
                 // Áp dụng phương thức kiểm tra và cập nhật trạng thái chữ ký
                 statusSignature = checkAndUpdateSignatureStatus(con, idImport, statusSignature);
+                statusSignature = checkChangeDatabase(con, userId, idImport, statusSignature);
 
 
                 User u = new UserDAO().selectById(idUser);
@@ -452,6 +466,7 @@ public class OrderDAO extends AbsDAO<Order>{
 
                 // Áp dụng phương thức kiểm tra và cập nhật trạng thái chữ ký
                 statusSignature = checkAndUpdateSignatureStatus(con, idImport, statusSignature);
+                statusSignature = checkChangeDatabase(con, userId, idImport, statusSignature);
 
 
                 User u = new UserDAO().selectById(idUser);
@@ -511,6 +526,7 @@ public class OrderDAO extends AbsDAO<Order>{
 
                 // Áp dụng phương thức kiểm tra và cập nhật trạng thái chữ ký
                 statusSignature = checkAndUpdateSignatureStatus(con, idImport, statusSignature);
+                statusSignature = checkChangeDatabase(con, userId, idImport, statusSignature);
 
                 User u = new UserDAO().selectById(idUser);
                 Payment pay = new PaymentDAO().selectById(idPayment);
@@ -527,6 +543,83 @@ public class OrderDAO extends AbsDAO<Order>{
 
         return orders;
     }
+
+    private int checkChangeDatabase(Connection con, int userId, int orderId, int currentStatus) throws SQLException {
+        // Nếu trạng thái đơn hàng là 6, bỏ qua kiểm tra
+        if (currentStatus == 6) {
+            return currentStatus;
+        }
+
+        try {
+            // Tính toán serialized_data
+            String serializedData = getSerializedDataForOrder(orderId); // Giả sử hàm này có hiệu quả
+            String calculatedHash = Hash.calculateHash(serializedData.getBytes(StandardCharsets.UTF_8));
+            System.out.println("Hash tính được: " + calculatedHash);
+
+            // Truy vấn để lấy chữ ký và khóa công khai
+            String sql = " SELECT os.signature, ku.public_key FROM order_signatures os LEFT JOIN key_user ku ON ku.user_id = ? AND ku.status = 'ON' WHERE os.order_id = ? ";
+
+            try (PreparedStatement st = con.prepareStatement(sql)) {
+                st.setInt(1, userId);
+                st.setInt(2, orderId);
+
+                try (ResultSet rs = st.executeQuery()) {
+                    if (rs.next()) {
+                        String signature = rs.getString("signature");
+                        String publicKey = rs.getString("public_key");
+
+                        // Nếu không có chữ ký
+                        if (signature == null) {
+                            // Nếu hash đúng trở lại, trạng thái = 1
+                            if (isHashCorrect(serializedData, calculatedHash)) {
+                                updateOrderStatus(con, orderId, 1);
+                                return 1;
+                            } else {
+                                // Hash không đúng -> trạng thái = 2
+                                updateOrderStatus(con, orderId, 2);
+                                return 2;
+                            }
+                        }
+
+                        // Nếu có chữ ký, giải mã hash từ chữ ký
+                        String decryptedHash = RSA.decryptHash(publicKey, signature);
+                        System.out.println("Hash giải mã: " + decryptedHash);
+
+                        // So sánh hash đã tính và hash giải mã
+                        if (calculatedHash.equals(decryptedHash)) {
+                            // Hash khớp -> trạng thái = 3
+                            updateOrderStatus(con, orderId, 3);
+                            return 3;
+                        } else {
+                            // Hash không khớp -> trạng thái = 2
+                            updateOrderStatus(con, orderId, 2);
+                            return 2;
+                        }
+                    }
+                }
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            // Nếu có lỗi, coi là trạng thái không hợp lệ
+            updateOrderStatus(con, orderId, 2);
+            return 2;
+        }
+
+        // Trả về trạng thái hiện tại nếu không có thay đổi
+        return currentStatus;
+    }
+
+    // Hàm kiểm tra hash có đúng hay không
+    private boolean isHashCorrect(String serializedData, String calculatedHash) throws NoSuchAlgorithmException {
+        // Tính lại hash từ dữ liệu gốc
+        String recalculatedHash = Hash.calculateHash(serializedData.getBytes(StandardCharsets.UTF_8));
+        System.out.println("Hash tính lại: " + recalculatedHash);
+
+        // So sánh với hash đã tính trước đó
+        return recalculatedHash.equals(calculatedHash);
+    }
+
+
 
     // Phương thức kiểm tra và cập nhật signature_status_id
     private int checkAndUpdateSignatureStatus(Connection con, int orderId, int currentStatusSignature) throws SQLException {
@@ -590,6 +683,74 @@ public class OrderDAO extends AbsDAO<Order>{
         // Không thay đổi trạng thái
         return currentStatusSignature;
     }
+
+    public String getSerializedDataForOrder(int orderId) throws SQLException {
+        StringBuilder serializedData = new StringBuilder();
+
+        // Câu lệnh SQL để lấy dữ liệu từ các bảng, chỉ lấy các trường cần thiết
+        String query = "SELECT " +
+                "u.user_id, " +
+                "u.email, " +
+                "o.order_id, " +
+                "o.total_price, " +
+                "o.booking_date, " +
+                "o.shipping_fee " +
+                "FROM orders o " +
+                "JOIN users u ON o.user_id = u.user_id " +
+                "WHERE o.order_id = ?";  // Sử dụng orderId từ tham số
+
+        try (Connection con = JDBCUtil.getConnection();
+             PreparedStatement ps = con.prepareStatement(query)) {
+            ps.setInt(1, orderId);  // Gán giá trị orderId vào câu lệnh SQL
+
+            try (ResultSet rs = ps.executeQuery()) {
+                if (rs.next()) {
+                    // Thêm thông tin user vào serializedData
+                    serializedData.append("user_id:").append(rs.getInt("user_id")).append(",");
+                    serializedData.append("email:").append(rs.getString("email")).append(",");
+                    serializedData.append("order_id:").append(rs.getInt("order_id")).append(",");
+                    serializedData.append("total_price:").append(rs.getDouble("total_price")).append(",");
+
+                    // Định dạng ngày giờ booking (bao gồm phần .0 sau giây)
+                    java.sql.Timestamp timestamp = rs.getTimestamp("booking_date");
+                    if (timestamp != null) {
+                        // Định dạng ngày giờ với .0 sau giây (yyyy-MM-dd HH:mm:ss.SSS)
+                        String formattedDate = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.S").format(timestamp);
+                        serializedData.append("booking_date:").append(formattedDate).append(",");
+                    } else {
+                        serializedData.append("booking_date:null,");
+                    }
+
+                    // Thêm phí vận chuyển
+                    serializedData.append("shipping_fee:").append(rs.getDouble("shipping_fee")).append(",");
+
+                    // Lấy thông tin chi tiết đơn hàng từ bảng OrderDetails
+                    String detailQuery = "SELECT p.product_id, p.product_name, od.price, od.quantity " +
+                            "FROM orderdetails od " +
+                            "JOIN products p ON od.product_id = p.product_id " +
+                            "WHERE od.order_id = ?";
+
+                    try (PreparedStatement psDetail = con.prepareStatement(detailQuery)) {
+                        psDetail.setInt(1, orderId);  // Gán giá trị orderId vào câu lệnh SQL
+
+                        try (ResultSet rsDetail = psDetail.executeQuery()) {
+                            while (rsDetail.next()) {
+                                serializedData.append("product_id:").append(rsDetail.getInt("product_id")).append(",");
+                                serializedData.append("product_name:").append(rsDetail.getString("product_name")).append(",");
+                                serializedData.append("price:").append(rsDetail.getDouble("price")).append(",");
+                                serializedData.append("quantity:").append(rsDetail.getInt("quantity")).append(",");
+                            }
+                        }
+                    }
+                }
+            }
+        }
+
+        // Xóa dấu phẩy cuối cùng để chuỗi không có dấu phẩy thừa
+
+        return serializedData.toString();
+    }
+
 
     // Hàm tiện ích để cập nhật trạng thái đơn hàng
     private void updateOrderStatus(Connection con, int orderId, int newStatus) throws SQLException {
@@ -978,14 +1139,32 @@ public class OrderDAO extends AbsDAO<Order>{
 
         return result;
     }
-    public static void main(String[] args) {
+    public static void main(String[] args) throws SQLException, NoSuchAlgorithmException {
         OrderDAO orderDAO = new OrderDAO();
-        orderDAO.updateStatusOrder(3,new StatusOrder(13));
-        double[]rs=orderDAO.revenueForWeek(Date.valueOf(LocalDateTime.now().toLocalDate()));
-        for (int i=0; i<5; i++){
-            System.out.println(rs[i]);
+        Order order = orderDAO.selectById(1);
+//        orderDAO.updateStatusOrder(3,new StatusOrder(13));
+//        double[]rs=orderDAO.revenueForWeek(Date.valueOf(LocalDateTime.now().toLocalDate()));
+//        for (int i=0; i<5; i++){
+//            System.out.println(rs[i]);
+//        }
+        UserDAO userDAO = new UserDAO();
+        User user = userDAO.selectById(4);
+        String hash = Hash.calculateHash(orderDAO.getSerializedDataForOrder(2).toString().getBytes(StandardCharsets.UTF_8));
+        System.out.println(hash);
+        // lay public key giai ma chu ky
+        KeyUserDAO keyUserDAO = new KeyUserDAO();
+        KeyUser keyUser = keyUserDAO.selectByUserIdStatus(user.getUserId(), "ON");
+        OrderSignatureDAO orderSignatureDao = new OrderSignatureDAO();
+        OrderSignature orderSignature = orderSignatureDao.selectByOrderId(order.getOrderId());
+        // Lấy hash đã lưu từ cơ sở dữ liệu
+        String storedHash = RSA.decryptHash(keyUser.getKey(), orderSignature.getSignature());
+//        String storedHash = orderSignatureDAO.getHashByOrderId(orderId);
+        System.out.println(storedHash);
+        if (hash.equals(storedHash)){
+            System.out.println("don hang khong bi thay doi");
+        }else {
+            System.out.println("Bi thay doi");
         }
-
     }
 
 }
