+++ src/main/java/database/OrderSignatureDAO.java
@@ -338,187 +338,6 @@ public class OrderSignatureDAO implements DAOInterface<OrderSignature> {
     }
 
 
-// Các phương thức còn lại trong lớp OrderSignatureDAO (selectAll, selectById, insert, update, delete...)
-
-
-    public static boolean validateOrderHash(int orderId, int userId) throws SQLException, NoSuchAlgorithmException {
-        // Lấy dữ liệu serialized cho đơn hàng từ phương thức đã viết
-        OrderSignatureDAO orderSignatureDAO = new OrderSignatureDAO();
-
-        String serializedData = orderSignatureDAO.getSerializedDataForOrder(orderId);
-
-        // Tính toán hash từ serializedData
-        String calculatedHash = Hash.calculateHash(serializedData.getBytes(StandardCharsets.UTF_8));
-        KeyUserDAO keyUserDAO = new KeyUserDAO();
-        KeyUser keyUser = keyUserDAO.selectByUserIdStatus(userId, "ON");
-        OrderSignatureDAO orderSignatureDao = new OrderSignatureDAO();
-        OrderSignature orderSignature = orderSignatureDao.selectByOrderId(orderId);
-        // Lấy hash đã lưu từ cơ sở dữ liệu
-        String storedHash = RSA.decryptHash(keyUser.getKey(), orderSignature.getSignature());
-//        String storedHash = orderSignatureDAO.getHashByOrderId(orderId);
-
-        // So sánh hash tính toán và hash đã lưu
-        return storedHash != null && storedHash.equals(calculatedHash);
-    }
-
-    public String getSerializedDataForOrder(int orderId) throws SQLException {
-        StringBuilder serializedData = new StringBuilder();
-
-        // Câu lệnh SQL để lấy dữ liệu từ các bảng, chỉ lấy các trường cần thiết
-        String query = "SELECT " +
-                "u.user_id, " +
-                "u.email, " +
-                "o.order_id, " +
-                "o.total_price, " +
-                "o.booking_date, " +
-                "o.shipping_fee " +
-                "FROM orders o " +
-                "JOIN users u ON o.user_id = u.user_id " +
-                "WHERE o.order_id = ?";  // Sử dụng orderId từ tham số
-
-        try (Connection con = JDBCUtil.getConnection();
-             PreparedStatement ps = con.prepareStatement(query)) {
-            ps.setInt(1, orderId);  // Gán giá trị orderId vào câu lệnh SQL
-
-            try (ResultSet rs = ps.executeQuery()) {
-                if (rs.next()) {
-                    // Thêm thông tin user vào serializedData
-                    serializedData.append("user_id:").append(rs.getInt("user_id")).append(",");
-                    serializedData.append("email:").append(rs.getString("email")).append(",");
-                    serializedData.append("order_id:").append(rs.getInt("order_id")).append(",");
-                    serializedData.append("total_price:").append(rs.getDouble("total_price")).append(",");
-
-                    // Định dạng ngày giờ booking (bao gồm phần .0 sau giây)
-                    java.sql.Timestamp timestamp = rs.getTimestamp("booking_date");
-                    if (timestamp != null) {
-                        // Định dạng ngày giờ với .0 sau giây (yyyy-MM-dd HH:mm:ss.SSS)
-                        String formattedDate = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss.S").format(timestamp);
-                        serializedData.append("booking_date:").append(formattedDate).append(",");
-                    } else {
-                        serializedData.append("booking_date:null,");
-                    }
-
-                    // Thêm phí vận chuyển
-                    serializedData.append("shipping_fee:").append(rs.getDouble("shipping_fee")).append(",");
-
-                    // Lấy thông tin chi tiết đơn hàng từ bảng OrderDetails
-                    String detailQuery = "SELECT p.product_id, p.product_name, od.price, od.quantity " +
-                            "FROM orderdetails od " +
-                            "JOIN products p ON od.product_id = p.product_id " +
-                            "WHERE od.order_id = ?";
-
-                    try (PreparedStatement psDetail = con.prepareStatement(detailQuery)) {
-                        psDetail.setInt(1, orderId);  // Gán giá trị orderId vào câu lệnh SQL
-
-                        try (ResultSet rsDetail = psDetail.executeQuery()) {
-                            while (rsDetail.next()) {
-                                serializedData.append("product_id:").append(rsDetail.getInt("product_id")).append(",");
-                                serializedData.append("product_name:").append(rsDetail.getString("product_name")).append(",");
-                                serializedData.append("price:").append(rsDetail.getDouble("price")).append(",");
-                                serializedData.append("quantity:").append(rsDetail.getInt("quantity")).append(",");
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        // Xóa dấu phẩy cuối cùng để chuỗi không có dấu phẩy thừa
-
-
-        return serializedData.toString();
-    }
-
-    public static List<Order> validateOrdersForUser(int userId) throws SQLException, NoSuchAlgorithmException {
-        // Khởi tạo đối tượng OrderSignatureDAO để sử dụng các phương thức cần thiết
-        OrderSignatureDAO orderSignatureDAO = new OrderSignatureDAO();
-
-        // Danh sách lưu trữ các đơn hàng bị thay đổi
-        List<Order> changedOrders = new ArrayList<>();
-
-        // Truy vấn tất cả các đơn hàng của người dùng này
-        String query = "SELECT order_id FROM orders WHERE user_id = ?";
-
-        try (Connection con = JDBCUtil.getConnection();
-             PreparedStatement ps = con.prepareStatement(query)) {
-
-            ps.setInt(1, userId);  // Gán giá trị userId vào câu lệnh SQL
-
-            try (ResultSet rs = ps.executeQuery()) {
-                // Duyệt qua từng đơn hàng của người dùng
-                while (rs.next()) {
-                    int orderId = rs.getInt("order_id");
-
-                    // Kiểm tra tính hợp lệ của hash cho từng đơn hàng
-                    boolean isValid = validateOrderHash(orderId, userId);
-
-                    if (!isValid) {
-                        // Nếu hash không hợp lệ, thông báo rằng đơn hàng đã bị thay đổi và lưu đơn hàng này vào danh sách
-                        System.out.println("Don hang co order_id = " + orderId + " da bi thay doi.");
-
-                        // Lấy thông tin đơn hàng bị thay đổi
-                        OrderDAO orderDAO = new OrderDAO();
-                        Order order = orderDAO.selectById(orderId);
-                        changedOrders.add(order);
-                    }
-                }
-
-                // Nếu không có đơn hàng nào bị thay đổi
-                if (changedOrders.isEmpty()) {
-                    System.out.println("Khong co don hang nao bi thay doi.");
-                }
-            }
-        } catch (SQLException | NoSuchAlgorithmException e) {
-            e.printStackTrace();  // In ra thông báo lỗi nếu có ngoại lệ
-            throw e;  // Ném lại ngoại lệ cho các lớp phía trên xử lý
-        }
-
-        // Trả về danh sách các đơn hàng bị thay đổi
-        return changedOrders;
-    }
-
-    // Main method for testing the validateOrderHash method
-    public static void main(String[] args) throws SQLException {
-        OrderSignatureDAO orderSignatureDAO = new OrderSignatureDAO();
-
-        int orderId = 1; // Example order ID to test with
-
-        String serializedData = orderSignatureDAO.getSerializedDataForOrder(orderId);
-        System.out.println(serializedData);
-
-        int userId = 4;  // Ví dụ userId cần kiểm tra
-        try {
-            // Validate the order hash for the specified orderId
-            boolean isValid = validateOrderHash(orderId, userId);
-
-            if (isValid) {
-                System.out.println("Đơn hàng ko bị thay đổi");
-            } else {
-                System.out.println("Đơn hàng bị thay đổi");
-            }
-        } catch (SQLException e) {
-            System.err.println("Database error: " + e.getMessage());
-        } catch (NoSuchAlgorithmException e) {
-            System.err.println("Hashing error: " + e.getMessage());
-        } catch (Exception e) {
-            System.err.println("Unexpected error: " + e.getMessage());
-        }
-
-
-
-        try {
-            // Kiểm tra các đơn hàng của người dùng và in ra thông báo nếu có đơn hàng bị thay đổi
-            validateOrdersForUser(userId);
-        } catch (SQLException e) {
-            System.err.println("Lỗi cơ sở dữ liệu: " + e.getMessage());
-        } catch (NoSuchAlgorithmException e) {
-            System.err.println("Lỗi trong quá trình tính toán hash: " + e.getMessage());
-        } catch (Exception e) {
-            System.err.println("Lỗi không xác định: " + e.getMessage());
-        }
-    }
 }
-
-
